fix: remove inputCheckboxLabel
