+++ packages/react/src/components/form/components/InputCheckbox/index.ts
@@ -1,4 +1,2 @@
 export * from './InputCheckbox';
-export { default as InputCheckboxLabel } from './InputCheckboxLabel';
-export * from './InputCheckboxLabel';
 export * from './components';
